Fix PTRACE_GETMAPFD argument.
Even though this is a file descriptor, it still has to be a 'long'
because it is written to via a ptrace() data argument, which expects
its argument to be a long.

+++ libproc/Psymtab.c
@@ -1084,7 +1084,7 @@ Pbuild_file_symtab(struct ps_prochandle *P, file_info_t *fptr)
 
 	GElf_Ehdr ehdr;
 
-	int fd;
+	long fd;
 	Elf_Data *shdata;
 	Elf_Scn *scn;
 	Elf *elf = NULL;
